+++ src/Connect-MS365.psm1
@@ -4,7 +4,10 @@ Connects to a given online service of Microsoft.
 
 .DESCRIPTION
 Connects to a given online service of Microsoft.
-One or multiple service names can be chosen.
+One or multiple service names can be chosen. Supports connection handling for
+- Microsoft Online (MSOL)
+- Exchange Online (EOL)
+- Teams
 
 .PARAMETER Service
 Specifies the service to connect to. May be a list of multiple services to use.
@@ -39,7 +42,7 @@ function Connect-MS365 {
         #service parameter to define to which services to connect to
         #are validated against available / implemented services
         [Parameter(Mandatory=$True, Position = 1)]
-        [ValidateSet('MSOL','EOL')]
+        [ValidateSet('MSOL','EOL','Teams')]
         [string[]]
         $Service,
         #mfa parameter if mfa authentication is necessary
@@ -47,6 +50,8 @@ function Connect-MS365 {
         [Parameter(Mandatory=$False, Position = 2, ParameterSetName = 'MFA')]
         [Switch]
         $MFA,
+        #Credential parameter to receive previously created PSCredential object.
+        #Primarily needed for testing calls 
         [Parameter(Mandatory = $False, Position = 3, ParameterSetName = 'Credential')]
         [PSCredential]
         $Credential
@@ -80,6 +85,16 @@ function Connect-MS365 {
                 }
                 else {
                     Connect-EOL -Credential $Credential
+                }
+                continue
+            }
+            # Exchange Online service
+            Teams {
+                if ($MFA) {
+                    Connect-Teams
+                }
+                else {
+                    Connect-Teams -Credential $Credential
                 }
                 continue
             }